Fixed directory structure, added element to /register, added test case for register
