+++ All Project Code/components/src/views/pages/register.ejs
