+++ All Project Code/components/src/views/partials/footer.ejs
