+++ All Project Code/components/src/views/partials/head.ejs
