+++ All Project Code/components/src/views/partials/navbar.ejs
